UnitTestLearning/UnitTestLearning/Controllers/ProductController.cs
using UnitTestLearning.Models;
using Microsoft.AspNetCore.Mvc;

namespace UnitTestLearning.Controllers
{
    public class ProductController : ControllerBase
    {
        List<Product> products = new List<Product>();

        public ProductController() { }

        public ProductController(List<Product> products)
        {
            this.products = products;
        }

        public IEnumerable<Product> GetAllProducts()
        {
            return products;
        }

        public async Task<IEnumerable<Product>> GetAllProductsAsync()
        {
            return await Task.FromResult(GetAllProducts());
        }

        public IActionResult GetProduct(int id)
        {
            var product = products.FirstOrDefault((p) => p.Id == id);
            if (product == null)
            {
                return NotFound();
            }
            return Ok(product);
        }

        public async Task<IActionResult> GetProductAsync(int id)
        {
            return await Task.FromResult(GetProduct(id));
        }
    }
}
